+++ src/frontend/src/pages/Login/Signup.tsx
@@ -1,7 +1,6 @@
-// pages/Signup.tsx
 import { useNavigate } from "react-router-dom";
 import { useFormFields } from "./hooks/useFormFields";
-import AuthForm from "./components/LoginComponent/AuthForm";
+import AuthForm from "./components/AuthForm";
 
 const Signup = () => {
   const navigate = useNavigate();
