+++ src/frontend/src/routes/router.tsx
@@ -3,6 +3,7 @@ import { createBrowserRouter } from "react-router-dom";
 //layouts
 import DefaultLayout from "@layouts/DefaultLayout";
 import LoginLayout from "@layouts/GuestLayout";
+
 import LoginPage from "@pages/Login/Login";
 import Signup from "@pages/Login/Signup";
